eximia_runtime/pyproject.toml
[project]
name = "eximia-runtime"
version = "0.1.0"
description = "Backend Runtime for eximIA.OS Multi-Agent System"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "litellm>=1.50.0",
    "pydantic>=2.0",
    "chromadb>=0.5.0",
    "redis>=5.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "typer>=0.12.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0",
    "markdown>=3.7",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pypdf>=3.15.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8.0",
]

[project.scripts]
eximia = "eximia_runtime.interfaces.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
include = ["*.py", "**/*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
